feat(api): parse attrs as liquid
