+++ libs/application-generic/package.json
@@ -95,7 +95,8 @@
     "rrule": "^2.7.2",
     "rxjs": "7.8.1",
     "sanitize-html": "^2.4.0",
-    "shortid": "^2.2.16"
+    "shortid": "^2.2.16",
+    "@maily-to/render": "^0.0.17"
   },
   "optionalDependencies": {
     "@novu/ee-shared-services": "workspace:*",
