+++ libs/application-generic/src/index.ts
@@ -17,3 +17,4 @@ export * from './tracing';
 export * from './usecases';
 export * from './utils';
 export * from './utils/inject-auth-providers';
+export * from './utils/process-node-attrs';
